+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [1.0.6] - 2020-12-26
 ### Added
 
 - GenerateDocumentationFile
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Follow ac-library v1.3
 - [Bug fix]Create all members in CreateOperatorCodeFixProvider
 - BinarySearch targets only IComparable<T>
+- Remove struct constraint from operator
 
 ## [1.0.5] - 2020-12-18
 ### Added